pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "SPICA"  # Or your chosen name (lowercase, no spaces)
version = "0.1.0"
authors = [
    {name="Adrià Casamitjana", email="adria.casamitjana@udg.edu"},
    {name="Miquel Gubau", email="gmgubau@gmail.com"},

]
description = "Automated evaluation of source dwell positions in Venezia™ lunar ovoids."
readme = "README.md"
license = "Apache-2.0"
license-files = ["LICENSE*"]
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Medical Science applications",

    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]
# List your core dependencies here instead of a separate requirements.txt
dependencies = [
    "numpy>=2.2.6",
    "scipy>=1.15.3",
    "matplotlib>=3.4.0", # For visualizing dwell positions
    "seaborn>=0.13.2",
    "scikit-image>=0.25.2",
    "scikit-learn>=1.7.2",
    "pillow>=12.1.1"
]

[project.urls]
"Homepage" = "https://github.com/acasamitjana/DwellVerify"
"Bug Tracker" = "https://github.com/acasamitjana/DwellVerify/issues"

[tool.setuptools.packages.find]
include = ["SPICA"]

[project.scripts]
# This allows users to type 'lunasentry' in their terminal to run your code
spica = "SPICA:app"